# swagger.yaml
openapi: 3.0.0
info:
  title: Practical skills assessment
  description: Node.js + MongoDB + Express
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development server
...
paths:
...
  /services:
    post:
      tags:
        - "Services"
      summary: Create a new service with its tiers
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Resume Writing"
                slug:
                  type: string
                  example: "resume-writing"
                description:
                  type: string
                  example: "This is resume writing"
                tiers:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Free"
                      price:
                        type: integer
                        example: 0
                      description:
                        type: string
                        example: "tier description"
              required:
                - name
                - slug
    get:
      tags:
        - "Services"
      summary: Get a all services
      description: Get a all services.
      operationId: getServices
  /services/{id}:
    parameters:
      - in: path
        name: id
        description: Service id
        required: true
        example: "6234d8f0d6d2e6f2c240ffd3"
        schema:
          type: string
    get:
      tags:
        - "Services"
      summary: Get a single service by id
      description: Get service by id.
      operationId: getServiceById
    put:
      tags:
        - "Services"
      summary: Update service
      operationId: updateService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Resume Writing"
                slug:
                  type: string
                  example: "resume-writing"
                description:
                  type: string
                  example: "This is resume writing"
                tiers:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        example: "Free"
                      price:
                        type: integer
                        example: 0
                      description:
                        type: string
                        example: "tier description"
              required:
                - name
                - slug
    delete:
      tags:
        - "Services"
      summary: Delete an service
      description: Delete service.
      operationId: deleteService
      responses:
        "200":
          description: Service deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Service deleted successfully"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /services/by-slug/{slug}:
    parameters:
      - in: path
        name: slug
        description: Slug of service
        required: true
        example: "resume-writing"
        schema:
          type: string
    get:
      tags:
        - "Services"
      summary: Get a single service by slug
      description: Get service by slug.
      operationId: getServiceBySlug
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"